// src/interfaces/IApgTngServiceOptions.ts
/** -----------------------------------------------------------------------
 * @module [Tng]
 * @author [APG] ANGELI Paolo Giusto
 * @version 0.9.0 [APG 2022/09/11] Beta
 * ------------------------------------------------------------------------
 */
export interface IApgTngServiceOptions {
  useCache?: boolean;
  cacheChunksLongerThan?: number;
  consoleLog?: boolean;
  deliverablesPath?: string;
}